requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
requests>=2.31.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
pyyaml>=6.0
aiohttp>=3.9.0
